인기 차트 데이터 호출 후 저장

+++ src/main/java/com/ltk/springproject/domain/Board.java
@@ -20,7 +20,7 @@ public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, updatable = false)
-    private Integer id; // INT(10) UNSIGNED
+    private Long id; // Integer -> Long으로 변경
 
     @Column(name = "regDate", nullable = false, updatable = false)
     private LocalDateTime regDate;
@@ -35,7 +35,7 @@ public class Board {
     private String name;
 
     @Column(name = "delStatus", nullable = false)
-    private int delStatus; // TINYINT(1) UNSIGNED, 0=정상, 1=삭제
+    private int delStatus;
 
     @Column(name = "delDate")
     private LocalDateTime delDate;
@@ -44,7 +44,7 @@ public class Board {
     protected void onCreate() {
         this.regDate = LocalDateTime.now();
         this.updateDate = LocalDateTime.now();
-        if (this.delStatus != 1) { // 기본값을 0으로 명시적으로 설정 (DB DEFAULT와 별개로 객체 생성 시)
+        if (this.delStatus != 1) {
             this.delStatus = 0;
         }
     }
